renamed onserr package to errors
