+++ errors/interface.go
@@ -1,4 +1,4 @@
-package onserr
+package errors
 
 // coder is an interface that allows you to 
 // extract a http status code from an error (or other object)
